# workshop-organizer/src/main/java/fr/oc/devops/backend/models/NotionModel.java
package fr.oc.devops.backend.models;

import jakarta.persistence.Entity;
import jakarta.persistence.Id;

@Entity
public class NotionModel {

//    @Id
//    @GeneratedValue(strategy = GenerationType.IDENTITY)
//    private Integer id;

    @Id
    private String name;

//    public Integer getId() {
//        return id;
//    }
//
//    public void setId(Integer id) {
//        this.id = id;
//    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}